implement review comments

+++ src/main/java/com/app/folioman/portfolio/models/response/PortfolioResponse.java
@@ -1,7 +1,8 @@
 package com.app.folioman.portfolio.models.response;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.List;
 
-public record PortfolioResponse(double totalPortfolioValue, List<PortfolioDetailsDTO> portfolioDetailsDTOS)
+public record PortfolioResponse(BigDecimal totalPortfolioValue, List<PortfolioDetailsDTO> portfolioDetailsDTOS)
         implements Serializable {}
